Refer to the Bruning (2019) paper
Add a reference to the Bruning (2019) paper in the README.  I found it
somewhat hard to find a detailed description of what the software
actually does (as opposed to how to use it).  The only mention of the
paper I could find was in the 0.5.0 release without a concrete citation,
I hope that I added the correct one.  Thanks for the software :)